pyproject.toml
[project]
name = "codeclone-iv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "psutil>=7.0.0",
    "tqdm>=4.67.1",
]
